lambda/layers/common_requirements.txt
PyJWT>=2.8.0
cryptography>=3.4.0
requests>=2.31.0